Add an evaluator
This adds an input which takes a numeric expression and attempts to run
it. If it succeeds it is placed in the value.

This allows you to perform basic arithmetic and then see the result.

This uses eval, which is scary, but hopefully it is somewhat safe
because of the limited acceptable input.